Generate pseudorandom dataframes with specific parameters

This function is a small utility to create a specific length dataframe with a set number of groups, specific mean/sd per group. Note that the total length of the dataframe will be n * n_grps.

generate_df(n = 10L, n_grps = 1L, with_seed = NULL, mean = c(10), sd = mean/10)

Arguments

n

An integer indicating the number of rows per group, default to 10

n_grps

An integer indicating the number of rows per group, defaults to 1

with_seed

A seed to make the randomization reproducible

mean

A number indicating the mean of the randomly generated values, must be a vector of equal length to the n_grps

sd

A number indicating the standard deviation of the randomly generated values, must be a vector of equal length to the n_grps

Value

a tibble/dataframe

Examples

generate_df()
#> # A tibble: 10 × 4
#>    row_id id    grp   values
#>     <int> <chr> <chr>  <dbl>
#>  1      1 Q10   grp-1   7.53
#>  2      2 Z08   grp-1  11.3 
#>  3      3 Q03   grp-1  10.2 
#>  4      4 X09   grp-1  11.1 
#>  5      5 R04   grp-1   9.65
#>  6      6 F07   grp-1   8.97
#>  7      7 J05   grp-1  10.1 
#>  8      8 Z05   grp-1  10.1 
#>  9      9 E01   grp-1  11.1 
#> 10     10 Y07   grp-1   9.92
generate_df(n = 100L, n_grps = 5L, with_seed = NULL, mean = seq(10, 50, length.out = 5))
#> # A tibble: 500 × 4
#>    row_id id    grp   values
#>     <int> <chr> <chr>  <dbl>
#>  1      1 Z073  grp-1  10.5 
#>  2      2 G049  grp-1   9.95
#>  3      3 X093  grp-1  10.1 
#>  4      4 F074  grp-1   9.27
#>  5      5 O008  grp-1  10.0 
#>  6      6 D023  grp-1  10.9 
#>  7      7 U082  grp-1   9.94
#>  8      8 Y082  grp-1  10.4 
#>  9      9 L055  grp-1   9.55
#> 10     10 U095  grp-1   9.31
#> # … with 490 more rows

library(dplyr)
#> Warning: package 'dplyr' was built under R version 4.2.1
#> 
#> Attaching package: 'dplyr'
#> The following objects are masked from 'package:stats':
#> 
#>     filter, lag
#> The following objects are masked from 'package:base':
#> 
#>     intersect, setdiff, setequal, union
generate_df(
  100L,
  n_grps = 5,
  mean = seq(10, 50, length.out = 5)
) |>
  group_by(grp) |>
  summarise(
    mean = mean(values), # mean is approx mean
    sd = sd(values),     # sd is approx sd
    n = n(),             # each grp is of length n
    # showing that the sd default of mean/10 works
    `mean/sd` = round(mean / sd, 1)
  )
#> # A tibble: 5 × 5
#>   grp    mean    sd     n `mean/sd`
#>   <chr> <dbl> <dbl> <int>     <dbl>
#> 1 grp-1  9.90 0.914   100      10.8
#> 2 grp-2 20.1  2.03    100       9.9
#> 3 grp-3 30.8  2.88    100      10.7
#> 4 grp-4 40.0  4.20    100       9.5
#> 5 grp-5 50.4  5.03    100      10