Generate pseudorandom dataframes with specific parameters

This function is a small utility to create a specific length dataframe with a set number of groups, specific mean/sd per group. Note that the total length of the dataframe will be n * n_grps.

Usage

generate_df(n = 10L, n_grps = 1L, with_seed = NULL, mean = c(10), sd = mean/10)

Arguments

n

An integer indicating the number of rows per group, default to 10

n_grps

An integer indicating the number of rows per group, defaults to 1

with_seed

A seed to make the randomization reproducible

mean

A number indicating the mean of the randomly generated values, must be a vector of equal length to the n_grps

sd

A number indicating the standard deviation of the randomly generated values, must be a vector of equal length to the n_grps

Value

a tibble/dataframe

Examples

generate_df()
#> # A tibble: 10 × 4
#>    row_id id    grp   values
#>     <int> <chr> <chr>  <dbl>
#>  1      1 H03   grp-1   9.32
#>  2      2 G08   grp-1  10.5 
#>  3      3 N08   grp-1   9.42
#>  4      4 D02   grp-1  10.7 
#>  5      5 I05   grp-1  10.8 
#>  6      6 Q10   grp-1  10.2 
#>  7      7 Y08   grp-1   9.51
#>  8      8 R05   grp-1  10.1 
#>  9      9 P05   grp-1  11.2 
#> 10     10 Z01   grp-1   8.83
generate_df(n = 100L, n_grps = 5L, with_seed = NULL, mean = seq(10, 50, length.out = 5))
#> # A tibble: 500 × 4
#>    row_id id    grp   values
#>     <int> <chr> <chr>  <dbl>
#>  1      1 U008  grp-1  10.5 
#>  2      2 D003  grp-1  10.4 
#>  3      3 F074  grp-1  10.2 
#>  4      4 U006  grp-1   8.22
#>  5      5 C010  grp-1  11.2 
#>  6      6 I021  grp-1   6.98
#>  7      7 T045  grp-1  10.4 
#>  8      8 B052  grp-1   9.69
#>  9      9 T085  grp-1   8.76
#> 10     10 U028  grp-1   9.39
#> # ℹ 490 more rows

library(dplyr)
#> 
#> Attaching package: ‘dplyr’
#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag
#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union
generate_df(
  100L,
  n_grps = 5,
  mean = seq(10, 50, length.out = 5)
) |>
  group_by(grp) |>
  summarise(
    mean = mean(values), # mean is approx mean
    sd = sd(values),     # sd is approx sd
    n = n(),             # each grp is of length n
    # showing that the sd default of mean/10 works
    `mean/sd` = round(mean / sd, 1)
  )
#> # A tibble: 5 × 5
#>   grp    mean    sd     n `mean/sd`
#>   <chr> <dbl> <dbl> <int>     <dbl>
#> 1 grp-1  10.1 0.997   100      10.1
#> 2 grp-2  19.9 2.13    100       9.3
#> 3 grp-3  30.3 2.73    100      11.1
#> 4 grp-4  40.3 4.10    100       9.8
#> 5 grp-5  49.7 4.73    100      10.5